Common - Creation-Deletion stats - back ok

+++ src/main/java/eu/heliumiot/console/jpa/interfaces/IHeliumDeviceChange.java
@@ -13,10 +13,10 @@ public class IHeliumDeviceChange {
 
     // ---
 
-    public IHeliumDeviceChange(String tenant_name, String tenant_uuid, int device_count) {
+    public IHeliumDeviceChange(String tenant_name, String tenant_uuid, Long device_count) {
         this.tenant_name = tenant_name;
         this.tenant_uuid = tenant_uuid;
-        this.device_count = device_count;
+        this.device_count = device_count.intValue();
     }
 
     // ---
